Fixed mouse_rel always returning 0

+++ src/input.rs
@@ -82,15 +82,15 @@ pub fn mouse_y() -> i32 {
 /// Gets the mouse relative position
 /// since the last internal application loop
 /// on the X axis.
-pub fn mouse_rel_x() -> u16 {
-    unsafe { dos_like_sys::mouserelx() as u16 }
+pub fn mouse_rel_x() -> i32 {
+    unsafe { dos_like_sys::mouserelx() }
 }
 
 /// Gets the mouse relative position
 /// since the last internal application loop
 /// on the Y axis.
-pub fn mouse_rel_y() -> u16 {
-    unsafe { dos_like_sys::mouserely() as u16 }
+pub fn mouse_rel_y() -> i32 {
+    unsafe { dos_like_sys::mouserely() }
 }
 
 impl KeyCode {
